Lagrangian Formalism for Electromagnetic Lagrangian
The report provides a comprehensive account of the Lagrangian formalism for Electrodynamics. The Lagrangian formalism for the massless electromagnetic fields is first introduced in section I, and the effects of photon mass are analyzed in section II. When a material enters its
superconducting state, many of its electromagnetic properties change. Section III investigates one such phenomenon, the Meissner effect, in which the material behaves like a diamagnet. Symmetries of the Lagrangian and the construction of the Stress-Energy tensor are discussed in section IV. This is followed by a discussion of Noether’s theorem in section V.
Section VI deals with the solution of the wave equation in covariant form constructed from Lorentz
invariant Green functions. Application of the Lagrangian method to quantum
electrodynamics and study of the behavior of the t’Hooft-Polyakov magnetic monopole is elaborated in
sections VII-IX. The final section discusses the construction of the path integral for a spinning particle and
establishes its classical dynamics to be of a massless point particle on the unit 2-sphere coupled to
a magnetic monopole field. Spin quantization is shown to emerge as a consequence of the global
gauge invariance of the spin path integral.